+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Tests for "Conway sequence".
 - Tests for "Dwarfs standing on the shoulders of giants".
+- Tests for "Mayan calculation".
 
 ## [1.1.0] - 2022-02-28
 ### Added